New Async classes. Conversion to Maven project
