+++ src/main/java/com/github/cjengineer18/desktopwindowtemplate/component/staticpanel/WaitingPanel.java
