+++ src/main/java/com/github/cjengineer18/desktopwindowtemplate/exception/UnavailableComponentException.java
