+++ src/main/resources/com/github/cjengineer18/desktopwindowtemplate/resources/bundle/panels.properties
