+++ src/main/resources/com/github/cjengineer18/desktopwindowtemplate/resources/bundle/panels_es.properties
